refactor(cosmo_val): fix sibling-module imports for package depth
Converting cosmo_val from a module to a package moved every file one level
deeper, so single-dot relative imports to top-level siblings now resolve
inside the package. Point them up a level:
- ..b_modes (core, pure_eb, cosebis)
- ..cosmology (core, pseudo_cl)
- ..rho_tau (psf_systematics, pseudo_cl)

Re-export cs_plots in __init__ from its origin (cs_util.plots) rather than
from core, which no longer needs the alias after the plotting methods moved
to their mixins. Preserves the sp_validation.cosmo_val.cs_plots attribute path.

Package now imports cleanly; test_cosmo_val + test_b_modes: 18 passed, 1 skipped.
